Replace single quotes wuith double quotes in javascript expressions involving acs_ListBulkActionClick
this is really needed only for the URL part, but was done on the name as well for consistency. The reason
is even a properly encoded URL could still contain single quotes, therefore breaking the expression,
while double quotes are expected to be encoded in valid, urlencoded URLs. This relies on trusting the developer
to properly encode URLs, which should be expected. An alternative could involve active escaping of
javascript special characters by e.g. ad_js_escape.