Add 'check_xotcl_xocc' test case, checking for 'xo::cc'.
The intention of this test is to expose the destruction of the '::xo::cc' object, as a side effect of the 'check_xotcl' test. The test will fail if executed after it, and succeed otherwise.

The object destruction affects other tests executed in the same batch, and breaks in particular xowf's 'create_folder_with_page'.

The 'check_xotcl' test runs the 'Serializer all' method, which is calling 'xo::at_cleanup', destroying the '::xo::cc' object.