Modernize ical support
Rewritten ical support based on NX instead of XOTcl, following
much closer to the the standards. The interface is compatible
with the old interface. Note that in case, applicatios have
subclassed/extended the xotcl-core ical support, it must be
as well ported to NX (which is an easy task).

Bumped verison number to 5.10.1d8