Provide facilities to validate against invalid SQL strings
We introduce a new page contract filter and nsf validator called "dbtext". They implement enforcing of a value to be useable in an SQL query. Currently, this means that the value should not contain the NUL character, but the definition may change in the future or become database-specific.
The html contract filter has also be extended to reject the NUL character.

The test suite has been updated/extended to reflect the changes.