Added parameter to define a default dbn to a database connection
By this change, one can now define a default dbn at the creation time
of a database connection object. Before, it was necessary to pass
the "-dbn" value to every single command. The parameter can still be
used for particular queries as before to overrule the default.

Example for defining a connection context to a pool named "legacy"
using the PostgreSQL database interface

   ::xo::db::DB-postgresql create ::xo::dc1 -dialect postgresql -dbn legacy
   lappend _ [::xo::dc1 get_value . {select count(*) from acs_objects}]
   lappend _ [::xo::dc get_value . {select count(*) from acs_objects}]

   #> 660 51606